impl: Remove Homepage duplicate from App.js
Now the hompage is contained to it's own file in Home.js. This should've
been completed in the original Homepage PBI

issue #80
