+++ frontend/flashfolio/src/App.js
@@ -13,6 +13,7 @@ import LogoutButton from "./LogoutButton.js"
 import UserInfoPreview from "./UserInfoPreview.js"
 import Navbar from "./Navbar.js";
 import NewDeckButton from "./NewDeckButton.js";
+import Home from "./Home.js";
 
 /* Styling */
 import "./App.css";
@@ -83,64 +84,6 @@ function App() {
 			</Switch>
 		</Router>
 		</loginContext.Provider>
-	);
-}
-
-/*
-Temporary Homepage
-
-Shown when the user loads the root directory
-
-TODO: When Hompage PBI is created, move this to it's own component.
-*/
-function Home() {
-
-	const { loginState } = useContext(loginContext);
-
-	const history = useHistory();
-	
-	const loadButton = () => {
-		history.push("/load");
-	  };
-
-	const profileButton = () => {
-		if (loginState === null)
-			history.push("/profile/");
-		else
-			history.push("/profile/" + loginState.googleId);
-	  };
-	
-
-	return (
-		<div class="container">
-			<div class="left">
-				<div class="logoCard">
-				<h1 class="logo">FLASH</h1>
-				</div>
-				<div class="slide">
-				<h1 class="logo">FOLIO</h1>
-				</div>
-			</div>
-			
-			<div class="right">
-				<div class="intro">
-					Hi! We're Flashfolio! A flashcard website you can use to study to your heart's desire.
-					If you would like to create a deck, please click "Log In." Otherwise, to peruse
-					our large variety of public decks, hit "Discover."
-				</div>
-				<div class ="buttons">
-					<button onClick={loadButton}>Discover</button>
-					{/*<button>Sign Up</button>*/}
-					{ loginState === null ?
-						<LoginButton /> :
-						<><LogoutButton />
-						<NewDeckButton />
-						</>}
-					{/*<UserInfoPreview />*/}
-					<button onClick={profileButton}>Profile</button>
-				</div>
-			</div>
-		</div>
 	);
 }
